Do not clear the prisma migrations table when clearing data
The idea is to make testing the reload of scryfall data simpler not to blow
the database away entirely.

+++ packages/database/src/reset-scryfall.ts
@@ -24,6 +24,10 @@ async function clearEntireDatabase() {
     WHERE schemaname = 'public'`;
 
   for (const {tablename} of result) {
+    if (tablename.startsWith('_')) {
+      console.log(`Skipping table ${tablename}`);
+      continue;
+    }
     await prisma.$executeRaw(
       Prisma.sql`TRUNCATE TABLE "${Prisma.raw(tablename)}" CASCADE`
     );
